Update 'package.json' with '"build": "vite build" (so that we can run 'npm run build'), and update 'purge' in 'tailwind.config.js', then run 'npm run build'
This generates the files in the 'dist' folder

Without updating 'purge' to say which files to look in, running 'npm run build' would generate ALL tailwind CSS classes, and make a huge CSS file in build.

By updating 'purge' to tell it where the CSS will be used, it will look in these files and only keep the necessary css.